Configure the `public_addr` ceph option
For now, we set it to the `SERVICE_HOST` value. I have a suspicion
this might not work for the cases where `SERVICE_HOST` is a floating
IP address rather than a fixed address (when running devstack on
OpenStack).

If it doesn't we'll have to add a support for a host IP address, but
that's a bigger change so let's try this first.